+++ controllers/main/state.py
@@ -1,0 +1,90 @@
+from math import sin, cos, pi
+
+class State:
+    def __init__(self, robot, timestep):
+        self.robot = robot
+        self.timestep = timestep
+        
+        # Initialize position state
+        self.x = 0.0  # X position in meters
+        self.y = 0.0  # Y position in meters
+        self.theta = 0.0  # Heading in radians
+        
+        # Get and enable wheel encoders
+        self.left_encoder = robot.getDevice('left_pos')
+        self.right_encoder = robot.getDevice('right_pos')
+        self.left_encoder.enable(timestep)
+        self.right_encoder.enable(timestep)
+        
+        # Robot physical parameters (in meters)
+        self.WHEEL_RADIUS = 0.08  # 8cm = 0.08m
+        self.WHEEL_BASE = 0.360975  # 36.195cm = 0.360975m (distance between wheels)
+        
+        # Store previous encoder values
+        self.prev_left_ticks = 0
+        self.prev_right_ticks = 0
+        
+        # Wait for first valid encoder readings
+        robot.step(timestep)
+        self.prev_left_ticks = self.left_encoder.getValue()
+        self.prev_right_ticks = self.right_encoder.getValue()
+
+    def update(self):
+        """Update the robot's estimated position based on wheel encoder readings"""
+        # Get current encoder values
+        left_ticks = self.left_encoder.getValue()
+        right_ticks = self.right_encoder.getValue()
+        
+        # Calculate change in encoder ticks
+        delta_left = left_ticks - self.prev_left_ticks
+        delta_right = right_ticks - self.prev_right_ticks
+        
+        # Apply scaling factor to compensate for systematic errors
+        TICK_SCALE = 0.98  # Adjust this value based on testing
+        delta_left *= TICK_SCALE
+        delta_right *= TICK_SCALE
+        
+        # Debug print when wheels move
+        if abs(delta_left) > 0.001 or abs(delta_right) > 0.001:
+            print(f"Wheel movement detected - Left: {delta_left:.4f}, Right: {delta_right:.4f}")
+        
+        # Convert ticks to distance traveled by each wheel
+        left_dist = delta_left * self.WHEEL_RADIUS
+        right_dist = delta_right * self.WHEEL_RADIUS
+        
+        # Calculate distance traveled by robot center
+        dist = (left_dist + right_dist) / 2.0
+        
+        # Calculate change in heading
+        delta_theta = (right_dist - left_dist) / self.WHEEL_BASE
+        
+        # Average heading during movement
+        avg_theta = self.theta + delta_theta / 2.0
+        
+        # Update position
+        self.x += dist * cos(avg_theta)
+        self.y += dist * sin(avg_theta)
+        self.theta = (self.theta + delta_theta) % (2 * pi)
+        
+        # Store current encoder values for next update
+        self.prev_left_ticks = left_ticks
+        self.prev_right_ticks = right_ticks
+    
+    def get_position(self):
+        """Return current estimated position and heading"""
+        return {
+            'x': self.x,
+            'y': self.y,
+            'theta': self.theta,
+            'theta_degrees': self.theta * 180 / pi
+        }
+    
+    def reset(self):
+        """Reset position estimation to origin"""
+        self.x = 0.0
+        self.y = 0.0
+        self.theta = 0.0
+        
+        # Reset encoder references
+        self.prev_left_ticks = self.left_encoder.getValue()
+        self.prev_right_ticks = self.right_encoder.getValue()
