+++ worlds/.Rectangle_Arena.wbproj
@@ -1,5 +1,5 @@
 Webots Project File version R2023b
-perspectives: 000000ff00000000fd0000000300000000000001ca0000039ffc0200000001fb00000012005300630065006e0065005400720065006501000000000000039f0000000000000000000000010000023a000002edfc0200000001fb0000001400540065007800740045006400690074006f00720100000016000002ed0000004100ffffff0000000300000780000000d8fc0100000002fb0000000e0043006f006e0073006f006c006501000001d0000005b00000000000000000fb0000001a0043006f006e0073006f006c00650041006c006c0041006c006c0100000000000007800000006900ffffff00000544000002ed00000004000000040000000100000008fc00000000
+perspectives: 000000ff00000000fd0000000300000000000001ca0000039ffc0200000001fb00000012005300630065006e0065005400720065006501000000000000039f000000000000000000000001000000930000032dfc0200000001fb0000001400540065007800740045006400690074006f007201000000160000032d0000004100ffffff000000030000078000000098fc0100000002fb0000000e0043006f006e0073006f006c006501000001d0000005b00000000000000000fb0000001a0043006f006e0073006f006c00650041006c006c0041006c006c0100000000000007800000006900ffffff000006eb0000032d00000004000000040000000100000008fc00000000
 simulationViewPerspectives: 000000ff00000001000000020000016c000004a00100000002010000000101
 sceneTreePerspectives: 000000ff00000001000000030000001e000000c0000000fa0100000002010000000201
 maximizedDockId: -1
